Introduce a config for RHTAP's TSSC samples
The motivation is to make it so the RHTAP sample templates do not
always pull in the latest rego policies from main branch of
ec-policies.

Because RHTAP users might not always be using the latest version of
the ec cli, there's a chance that an update to the policies might
not be backwards compatible, which could break RHTAP useres. So
that's why it's better if RHTAP users are not using main branch
ec-policies.

In future we might want to provide more than one of these, e.g. a
stable config for RHTAP 1.0 and for RHTAP 1.2. For that, let's
follow the example set by the `tssc-sample-templates` repo. If they
create different branches for different RHTAP releases, then we can
follow suit. For now let's start with just one config file.

Ref: https://issues.redhat.com/browse/EC-605