Add config for use with RHTAP Jenkins pipelines
We're currently working on Jenkins pipeline support for RHTAP 1.2.

Let's provide a config that can be used for images built in that
pipeline. As mentioned in the comments, there are no rules currently
in the specified collection, but we'll aim to add some soon.

There are some fresh new rules in the rhtap-jenkins collection now
that will be used by the config, see
https://github.com/enterprise-contract/ec-policies/pull/1091

Ref: https://issues.redhat.com/browse/EC-746