Pin the policies refs used in rhtap-jenkins config
There was a new rego function added recently to ec called
`ec.oci.image_files`. The main branch of ec-policies now uses that
function and hence causes problems for older versions of ec where
that function is not defined.

To avoid breaking RHTAP Jenkins users, pin the policy refs so that
the the `ec.oci.image_file` function is not encountered.

Note:
- We can't use the release-v0.4 branch here, even though it's likely
  RHTAP users will initially have ec-v0.4 because the Jenkins rego
  rules were added after that branch was created.
- However, ec-v0.5 should be compatible with the release-v0.5
  ec-policies rego. Actually it's expected that RHTAP users will
  upgrade to RHTAS 1.1.0 when it is released, which does include
  ec-v0.5.
- So that's why it should be fine to use the release-v0.5 branch,
  even though it's a little newer version that the ec-v0.4 binary
  likely being used.