Include rpm task list source in redhat-rpms config
The rpm-specific tasks created for the Konflux rpm build pipelines
are not defined in the build-definitions repo, so they're not
present in the trusted task list in the "acceptable bundles" bundle
created from that repo.

Recently a new bundle was created for the rpm tasks trusted task
list, so let's include that in the "redhat-rpms" sample config here
in this repo.

Followup for:
Ref: https://issues.redhat.com/browse/EC-1203