Add new stable configs for ec v0.7
As per usual, we create these configs when a new release branch
(i.e. Red Hat official) build of Conforma CLI is created, as a
convenient way to provide a reference to a policy library that will
remain compatible with Conforma v0.7.

Since RHADS-SCC/RHTAP users are likely not using the latest main
branch builds, this provides better stability compared to using the
latest main branch of the policy repo.

Ref: https://issues.redhat.com/browse/EC-1384